admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["FabrizioPerri <fabrizioperri03@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.3"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-session = "^0.8.0"
bcrypt = "^4.2.0"
flask-bcrypt = "^1.0.1"
pydantic = "^2.9.2"
email-validator = "^2.2.0"
minio = "^7.2.9"
flask-cors = "^5.0.0"
marshmallow = "^3.23.1"
matplotlib = "^3.9.2"
pyopenssl = "^24.2.1"
authlib = "^1.3.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
